Move V1 specific methods from AbstractTokenExchangeProvider closes #39988
